# src/weather_api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Weather API",
        "description": "API para consultar el clima y gestionar usuarios",
        "contact": {
            "email": "tuemail@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Obtener Usuarios",
                "description": "Retorna una lista de los usuarios",
                "operationId": "GetUsers",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios obtenida con éxito"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Actualizar datos del usuario",
                "description": "Actualiza el nombre y correo electrónico del usuario",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Updated"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johnupdated@example.com"
                                    },
                                    "language": {
                                        "type": "string",
                                        "example": "es"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado correctamente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Eliminar usuario",
                "description": "Elimina un usuario del sistema",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado correctamente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/weather/{city}": {
            "get": {
                "tags": [
                    "Weather"
                ],
                "summary": "Obtener clima de una ciudad",
                "description": "Retorna la temperatura, humedad y otros datos del clima",
                "operationId": "getWeather",
                "parameters": [
                    {
                        "name": "city",
                        "in": "path",
                        "description": "Nombre de la ciudad",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clima obtenido con éxito"
                    },
                    "400": {
                        "description": "Error en la solicitud"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/weather-history": {
            "get": {
                "tags": [
                    "Histories"
                ],
                "summary": "Obtener historial de ciudades",
                "description": "Retorna una lista del historial de ciudades del usuario",
                "operationId": "getHistoryCities",
                "responses": {
                    "200": {
                        "description": "Lista de historial de ciudades obtenida con éxito"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}